batch/pyproject.toml
[project]
name = "batch"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "lxml>=5.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
]
